pyproject.toml
[tool.poetry]
name = "herald"
version = "0.1.0"
description = "Automatically run tasks based on git status"
homepage = "https://github.com/rdpfeffer/herald"
repository = "https://github.com/rdpfeffer/herald"
authors = ["Ryan Pfeffer <rdpfeffer@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
attrs = "^19.1"
jsonschema = "^3.0"
invoke = "^1.3"
click = "^7.0"

[tool.poetry.dev-dependencies]
pylint = ">=2.3"
isort = "^4.3"
pytest = "^4.6"
pytest-cov = "^2.7"
mypy = "^0.720.0"
flake8 = "^3.7"
tox = "^3.13"
black = { version = "^19.3b0", allows-prereleases = true, python = "^3.6"}

[tool.poetry.scripts]
herald = 'herald.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

